Split /localreview and /localaddress out of /localgremlin
Factor the review-code and address-code stage bodies into skills/localgremlin/_core.py
and expose them as two foreground skills, /localreview and /localaddress, that share
the exact code paths the gremlin's review and address stages run. /localgremlin keeps
its current end-to-end behavior by calling the same helpers directly.

Closes #39
